svg/path/property/inheritance.tentative.svg

<svg xmlns="http://www.w3.org/2000/svg" xmlns:h="http://www.w3.org/1999/xhtml"><title>Inheritance of path properties</title><metadata><h:link href="https://svgwg.org/svg2-draft/paths.html#PathLengthProperty" rel="help"/><h:meta content="path-length does not inherit." name="assert"/></metadata><h:script src="/resources/testharness.js"/><h:script src="/resources/testharnessreport.js"/><h:script src="/css/support/inheritance-testcommon.js"/><script><![CDATA[

assert_not_inherited('path-length', 'none', '10');

  ]]></script></svg>
